+++ src/Core/ApiClient.php
@@ -5,9 +5,11 @@ declare(strict_types=1);
 namespace Bitrix24\SDK\Core;
 
 use Bitrix24\SDK\Core\Contracts\ApiClientInterface;
+use Bitrix24\SDK\Core\Credentials\Credentials;
 use Bitrix24\SDK\Core\Exceptions\InvalidArgumentException;
 use Bitrix24\SDK\Core\Exceptions\TransportException;
 use Bitrix24\SDK\Core\Response\DTO\RenewedAccessToken;
+use Bitrix24\SDK\Infrastructure\HttpClient\RequestId\RequestIdGeneratorInterface;
 use Fig\Http\Message\StatusCodeInterface;
 use Psr\Log\LoggerInterface;
 use Symfony\Contracts\HttpClient\Exception\TransportExceptionInterface;
@@ -18,7 +20,8 @@ class ApiClient implements ApiClientInterface
 {
     protected HttpClientInterface $client;
     protected LoggerInterface $logger;
-    protected Credentials\Credentials $credentials;
+    protected Credentials $credentials;
+    protected RequestIdGeneratorInterface $requestIdGenerator;
     /**
      * @const string
      */
@@ -32,14 +35,20 @@ class ApiClient implements ApiClientInterface
     /**
      * ApiClient constructor.
      *
-     * @param Credentials\Credentials $credentials
+     * @param Credentials $credentials
      * @param HttpClientInterface $client
+     * @param RequestIdGeneratorInterface $requestIdGenerator
      * @param LoggerInterface $logger
      */
-    public function __construct(Credentials\Credentials $credentials, HttpClientInterface $client, LoggerInterface $logger)
+    public function __construct(
+        Credentials                 $credentials,
+        HttpClientInterface         $client,
+        RequestIdGeneratorInterface $requestIdGenerator,
+        LoggerInterface             $logger)
     {
         $this->credentials = $credentials;
         $this->client = $client;
+        $this->requestIdGenerator = $requestIdGenerator;
         $this->logger = $logger;
         $this->logger->debug(
             'ApiClient.init',
@@ -64,9 +73,9 @@ class ApiClient implements ApiClientInterface
     }
 
     /**
-     * @return Credentials\Credentials
+     * @return Credentials
      */
-    public function getCredentials(): Credentials\Credentials
+    public function getCredentials(): Credentials
     {
         return $this->credentials;
     }
@@ -80,7 +89,10 @@ class ApiClient implements ApiClientInterface
      */
     public function getNewAccessToken(): RenewedAccessToken
     {
-        $this->logger->debug('getNewAccessToken.start');
+        $requestId = $this->requestIdGenerator->getRequestId();
+        $this->logger->debug('getNewAccessToken.start', [
+            'requestId' => $requestId
+        ]);
         if ($this->getCredentials()->getApplicationProfile() === null) {
             throw new InvalidArgumentException('application profile not set');
         }
@@ -103,14 +115,21 @@ class ApiClient implements ApiClientInterface
         );
 
         $requestOptions = [
-            'headers' => $this->getDefaultHeaders(),
+            'headers' => array_merge(
+                $this->getDefaultHeaders(),
+                [
+                    $this->requestIdGenerator->getHeaderFieldName() => $requestId
+                ]
+            ),
         ];
         $response = $this->client->request($method, $url, $requestOptions);
         $responseData = $response->toArray(false);
         if ($response->getStatusCode() === StatusCodeInterface::STATUS_OK) {
             $newAccessToken = RenewedAccessToken::initFromArray($responseData);
 
-            $this->logger->debug('getNewAccessToken.finish');
+            $this->logger->debug('getNewAccessToken.finish', [
+                'requestId' => $requestId
+            ]);
             return $newAccessToken;
         }
         if ($response->getStatusCode() === StatusCodeInterface::STATUS_BAD_REQUEST) {
@@ -129,12 +148,14 @@ class ApiClient implements ApiClientInterface
      */
     public function getResponse(string $apiMethod, array $parameters = []): ResponseInterface
     {
+        $requestId = $this->requestIdGenerator->getRequestId();
         $this->logger->info(
             'getResponse.start',
             [
                 'apiMethod' => $apiMethod,
                 'domainUrl' => $this->credentials->getDomainUrl(),
                 'parameters' => $parameters,
+                'requestId' => $requestId
             ]
         );
 
@@ -150,9 +171,15 @@ class ApiClient implements ApiClientInterface
             $parameters['auth'] = $this->getCredentials()->getAccessToken()->getAccessToken();
         }
 
+
         $requestOptions = [
             'json' => $parameters,
-            'headers' => $this->getDefaultHeaders(),
+            'headers' => array_merge(
+                $this->getDefaultHeaders(),
+                [
+                    $this->requestIdGenerator->getHeaderFieldName() => $requestId
+                ]
+            ),
             // disable redirects, try to catch portal change domain name event
             'max_redirects' => 0,
         ];
@@ -163,6 +190,7 @@ class ApiClient implements ApiClientInterface
             [
                 'apiMethod' => $apiMethod,
                 'responseInfo' => $response->getInfo(),
+                'requestId' => $requestId
             ]
         );